+++ app.js
@@ -5,8 +5,7 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-
+const mealsRouter = require('./routes/meals.route')
 var app = express();
 
 // view engine setup
@@ -20,7 +19,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
+app.use('/meals', mealsRouter)
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
