+++ routes/meals.route.js
@@ -4,6 +4,6 @@ const mealController = require('../controllers/meals.controller')
 
 // peticion -> middleware -> mealController.listMeals -> error;
 /* GET home page. */
-router.get('/', mealController.listMeals);
+router.get('/detail/:id', mealController.getProduct);
 
 module.exports = router;
