+++ views/detalleMenu.ejs
@@ -14,17 +14,17 @@
       <!-- Menu Section -->
       <div class="ps-row ps-padding-64" id="menu">
         <div class="ps-col l6 ps-padding-large">
-          <h1 class="ps-center">Carpaccio fresco</h1>
+          <h1 class="ps-center"><%= meal.name %> </h1>
           <br />
-          <h4>Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?</h4>
+          <h4><%= meal.description || 'Platillo Tipico' %> </h4>
           <p class="ps-text-grey">
-            U$S 65.50
+            <%= `${meal.currency} ${meal.price}` %>
           </p>
         </div>
 
         <div class="ps-col l6 ps-padding-large ps-img-menu">
           <img
-            src="/images/menu.jpg"
+            src=<%= meal.img %>
             class="ps-round ps-image ps-opacity-min"
             alt="Menu"
           />
