+++ views/error.ejs
@@ -1,3 +1,5 @@
-<h1><%= message %></h1>
+
+
+<h1>Ocurrio un error inesperado <%= message %></h1>
 <h2><%= error.status %></h2>
 <pre><%= error.stack %></pre>
